refactor: move serialization into Mirage.Serialization
Move reader and writer to a Mirage.Serialization namespace

BREAKING CHANGE: NetworkReader and NetworkWriter moved to Mirage.Serialization namespace
