+++ Assets/Mirage/Components/NetworkAnimator.cs
@@ -2,6 +2,7 @@ using System.Linq;
 using UnityEngine;
 using UnityEngine.Serialization;
 using Mirage.Logging;
+using Mirage.Serialization;
 
 namespace Mirage
 {
