+++ Assets/Mirage/Runtime/ClientObjectManager.cs
@@ -6,6 +6,7 @@ using Mirage.RemoteCalls;
 using UnityEngine;
 using UnityEngine.Serialization;
 using Mirage.Logging;
+using Mirage.Serialization;
 
 namespace Mirage
 {
