+++ Assets/Mirage/Runtime/Collections/SyncDictionary.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using JetBrains.Annotations;
+using Mirage.Serialization;
 
 namespace Mirage.Collections
 {
