+++ Assets/Mirage/Runtime/Collections/SyncList.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using Mirage.Serialization;
 
 namespace Mirage.Collections
 {
