+++ Assets/Mirage/Runtime/GameObjectSyncvar.cs
@@ -1,4 +1,5 @@
-﻿using UnityEngine;
+﻿using Mirage.Serialization;
+using UnityEngine;
 
 namespace Mirage
 {
