+++ Assets/Mirage/Runtime/NetworkIdentity.cs
@@ -6,6 +6,7 @@ using UnityEngine;
 using UnityEngine.Serialization;
 using UnityEngine.Events;
 using Mirage.Logging;
+using Mirage.Serialization;
 #if UNITY_EDITOR
 using UnityEditor;
 #if UNITY_2018_3_OR_NEWER
