+++ Assets/Mirage/Runtime/NetworkIdentitySyncvar.cs
@@ -1,4 +1,6 @@
-﻿namespace Mirage
+﻿using Mirage.Serialization;
+
+namespace Mirage
 {
 
     /// <summary>
