+++ Assets/Mirage/Runtime/RemoteCallHelper.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using Cysharp.Threading.Tasks;
 using UnityEngine;
 using Mirage.Logging;
+using Mirage.Serialization;
 
 namespace Mirage.RemoteCalls
 {
