+++ Assets/Mirage/Runtime/Serialization.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 999c5520b3d65450d8bcfba1231ec904
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
