+++ Assets/Mirage/Runtime/Serialization/NetworkReader.cs
@@ -13,7 +13,7 @@ using System.Text;
 using UnityEngine;
 using Mirage.Logging;
 
-namespace Mirage
+namespace Mirage.Serialization
 {
     /// <summary>
     /// a class that holds readers for the different types
@@ -25,20 +25,6 @@ namespace Mirage
     public static class Reader<T>
     {
         public static Func<NetworkReader, T> Read { internal get; set; }
-    }
-
-    /// <summary>
-    /// An object that implements this interface can find objects by their net id
-    /// This is used by readers when trying to deserialize gameobjects
-    /// </summary>
-    public interface IObjectLocator
-    {
-        /// <summary>
-        /// Finds a network identity by id
-        /// </summary>
-        /// <param name="netId">the id of the object to find</param>
-        /// <returns>The NetworkIdentity matching the netid or null if none is found</returns>
-        NetworkIdentity this[uint netId] { get; }
     }
 
     // Note: This class is intended to be extremely pedantic, and
