+++ Assets/Mirage/Runtime/Serialization/NetworkReaderPool.cs.meta
