+++ Assets/Mirage/Runtime/Serialization/NetworkWriter.cs.meta
