+++ Assets/Mirage/Runtime/Serialization/NetworkWriterPool.cs.meta
