+++ Assets/Mirage/Runtime/Serialization/TypeConversion.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8ab6c1ed99802413b9d356645175e225
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
