+++ Assets/Tests/Editor/NetworkIdentityCallbackTests.cs
@@ -6,6 +6,7 @@ using UnityEngine;
 using UnityEngine.Events;
 using static Mirage.Tests.LocalConnections;
 using Object = UnityEngine.Object;
+using Mirage.Serialization;
 
 namespace Mirage
 {
