+++ Assets/Tests/Editor/NetworkTransformTest.cs
@@ -1,5 +1,6 @@
 using NUnit.Framework;
 using UnityEngine;
+using Mirage.Serialization;
 
 namespace Mirage
 {
